file rename and restructure the directory
