+++ middlewares/jwt.js
@@ -13,7 +13,7 @@ const authenticate = jwt({
       // console.log(req.headers.authorization);
       return req.headers.authorization.split(" ")[1];
     } else if (req.query && req.query.token) {
-      return req.query.token;
+      return req.query.token.split(" ")[1];
     }
     return null;
   },
