+++ public/views/booking/booking.ejs
@@ -345,5 +345,5 @@
             </main>
             <%- include('../partials/bottom-a') -%>
 
-                <script src="/../../js/booking.js"></script>                
+                <script src="../../js/booking.js"></script>                
                 <%- include('../partials/bottom-b') -%>
